+++ src/main/java/dev/IESFranciscodelosRios/Controller/HubController.java
@@ -1,0 +1,4 @@
+package dev.IESFranciscodelosRios.Controller;
+
+public class HubController {
+}
