+++ src/main/java/dev/IESFranciscodelosRios/Domain/DAO/UserDAO.java
@@ -1,0 +1,4 @@
+package dev.IESFranciscodelosRios.Domain.DAO;
+
+public class UserDAO {
+}
