+++ src/main/java/dev/IESFranciscodelosRios/Domain/Model/Chat.java
@@ -1,0 +1,4 @@
+package dev.IESFranciscodelosRios.Domain.Model;
+
+public class Chat {
+}
