+++ src/main/java/dev/IESFranciscodelosRios/Thread/ChatThreadUpdate.java
@@ -1,0 +1,4 @@
+package dev.IESFranciscodelosRios.Thread;
+
+public class ChatThreadUpdate {
+}
