+++ src/main/java/dev/IESFranciscodelosRios/Utils/XMLManager.java
@@ -1,0 +1,4 @@
+package dev.IESFranciscodelosRios.Utils;
+
+public class XMLManager {
+}
