+++ src/main/java/module-info.java
@@ -2,8 +2,8 @@ module org.example {
     requires javafx.controls;
     requires javafx.fxml;
 
-    opens org.example to javafx.fxml;
-    exports org.example;
-    exports org.example.Controller;
-    opens org.example.Controller to javafx.fxml;
+    opens dev.IESFranciscodelosRios to javafx.fxml;
+    exports dev.IESFranciscodelosRios;
+    exports dev.IESFranciscodelosRios.Controller;
+    opens dev.IESFranciscodelosRios.Controller to javafx.fxml;
 }
