--- src/main/java/org/example/Controller/ChatRoomController.java
@@ -1,4 +1,0 @@
-package org.example.Controller;
-
-public class ChatRoomController {
-}
