--- src/main/java/org/example/Controller/LobbyController.java
@@ -1,4 +1,0 @@
-package org.example.Controller;
-
-public class LobbyController {
-}
