--- src/main/resources/org/example/Lobby.fxml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.text.Font?>
-
-
-<AnchorPane prefHeight="600.0" prefWidth="800.0" style="-fx-background-color: rgb(24, 111, 178);" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.Controller.LobbyController">
-   <children>
-      <Pane layoutX="89.0" layoutY="100.0" prefHeight="459.0" prefWidth="622.0" style="-fx-background-color: rgb(16, 124, 16);">
-         <children>
-            <TableView layoutX="64.0" layoutY="57.0" prefHeight="250.0" prefWidth="495.0">
-               <columns>
-                  <TableColumn prefWidth="248.79998779296875" text="C1" />
-                  <TableColumn minWidth="0.0" prefWidth="246.4000244140625" text="C2" />
-               </columns>
-            </TableView>
-            <Button layoutX="121.0" layoutY="362.0" mnemonicParsing="false" prefHeight="33.0" prefWidth="107.0" style="-fx-background-color: grey;" text="Enter" textFill="WHITE" />
-            <Button layoutX="390.0" layoutY="362.0" mnemonicParsing="false" prefHeight="33.0" prefWidth="107.0" style="-fx-background-color: grey;" text="Create Room" textFill="WHITE" />
-         </children>
-      </Pane>
-      <Label alignment="CENTER" layoutX="89.0" layoutY="26.0" prefHeight="65.0" prefWidth="622.0" style="-fx-background-color: grey;" text="LOBBY" textFill="WHITE">
-         <font>
-            <Font size="44.0" />
-         </font>
-      </Label>
-   </children>
-</AnchorPane>
